Fix Kafka broker settings
Previously, Kafka was configured with two advertised listeners - one on
host 'kafka' and other on 'localhost'.

On digging deeper, it was found that bootstrap server (same as broker in
this case) forwards each of these as metadata to clients. Clients
connect to one of these. When client attempts connection to localhost,
the attempt fails becuase in client container, 'localhost' resolves to a
different machine.

The fix is simple, advertise only one listener - on host 'kafka'.

Other: Set producer config to timeout on 10s (default was 60s which is
too long).