Fix UDP feed processing logic
The logic before assumed that each packet contained only one message.
This is not the case.

Before, the parser would extract one message frpom each packet and then
discard the reminaing packet.

Now, the parser keeps extracting messages from packet until the buffer
is over. While doing so it is safely assumed that:
- packets contains whole message (i.e. message is not distributed across
  packets);
- packets do not contain any stray bytes that belong to no message.